+++ example/lib/main.dart
@@ -36,12 +36,12 @@ class Home extends StatelessWidget {
             RaisedButton(
               child: Text('Open Second Page'),
               onPressed: () async {
-                final response = await Routes.sailor.navigate<bool>(
-                  "/secondPage",
-                  transitions: [
-                    SailorTransition.slide_from_top,
-                  ],
-                );
+                final response = await Routes.sailor
+                    .navigate<bool>("/secondPage", transitions: [
+                  SailorTransition.slide_from_top,
+                ], params: {
+                  'id': 123,
+                });
 
                 print("Response from SecondPage: $response");
               },
@@ -101,6 +101,7 @@ class SecondPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final args = Sailor.args<SecondPageArgs>(context);
+    final id = Sailor.param<int>(context, 'id');
 
     return Scaffold(
       appBar: AppBar(
@@ -111,6 +112,7 @@ class SecondPage extends StatelessWidget {
           mainAxisSize: MainAxisSize.min,
           children: <Widget>[
             Text(args?.text ?? 'Second Page'),
+            Text("Param('id'): $id"),
             RaisedButton(
               child: Text('Close Page'),
               onPressed: () {
@@ -233,8 +235,14 @@ class Routes {
     sailor.addRoutes([
       SailorRoute(
         name: "/secondPage",
-        builder: (context, args) => SecondPage(),
+        builder: (context, args, params) => SecondPage(),
         defaultArgs: SecondPageArgs('From default arguments!'),
+        params: [
+          SailorParam(
+            name: 'id',
+            defaultValue: 1234,
+          ),
+        ],
         defaultTransitions: [
           SailorTransition.slide_from_bottom,
           SailorTransition.zoom_in,
@@ -242,12 +250,12 @@ class Routes {
       ),
       SailorRoute(
         name: "/thirdPage",
-        builder: (context, args) => ThirdPage(),
+        builder: (context, args, params) => ThirdPage(),
         defaultTransitions: [SailorTransition.slide_from_left],
       ),
       SailorRoute(
         name: "/pushReplacePage",
-        builder: (context, args) => PushReplacePage(),
+        builder: (context, args, params) => PushReplacePage(),
       ),
     ]);
   }