+++ lib/sailor.dart
@@ -2,6 +2,7 @@ library sailor;
 
 export 'package:sailor/src/sailor.dart';
 export 'package:sailor/src/models/sailor_route.dart';
+export 'package:sailor/src/models/sailor_param.dart';
 export 'package:sailor/src/models/base_arguments.dart';
 export 'package:sailor/src/navigator_observers/sailor_logging_observer.dart';
 export 'package:sailor/src/models/sailor_options.dart';