+++ lib/src/models/arguments_wrapper.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/animation.dart';
-import 'package:sailor/sailor.dart';
+import 'package:sailor/src/models/base_arguments.dart';
+import 'package:sailor/src/models/sailor_param.dart';
 import 'package:sailor/src/transitions/sailor_transition.dart';
 
 class ArgumentsWrapper {
@@ -7,19 +8,48 @@ class ArgumentsWrapper {
   final List<SailorTransition> transitions;
   final Duration transitionDuration;
   final Curve transitionCurve;
+  final Map<String, dynamic> params;
+
+  /// Route params that are initially defined for the Route.
+  /// Note: These are needed here, because user must be able to retrieve
+  /// params from within the launched route. Since the only thing available
+  /// from inside a launched route are arguments, these paramters are provided
+  /// along with the arguments.
+  final Map<String, SailorParam> routeParams;
 
   ArgumentsWrapper({
     this.baseArguments,
     this.transitions,
     this.transitionDuration,
     this.transitionCurve,
+    this.params,
+    this.routeParams,
   });
+
+  ArgumentsWrapper copyWith({
+    BaseArguments baseArguments,
+    List<SailorTransition> transitions,
+    Duration transitionDuration,
+    Curve transitionCurve,
+    List<SailorParam> params,
+    Map<String, SailorParam> routeParams,
+  }) {
+    return ArgumentsWrapper(
+      baseArguments: baseArguments ?? this.baseArguments,
+      transitions: transitions ?? this.transitions,
+      transitionDuration: transitionDuration ?? this.transitionDuration,
+      transitionCurve: transitionCurve ?? this.transitionCurve,
+      params: params ?? this.params,
+      routeParams: routeParams ?? this.routeParams,
+    );
+  }
 
   @override
   String toString() {
     return 'ArgumentsWrapper{baseArguments: $baseArguments, '
         'transitions: $transitions, '
         'transitionDuration: $transitionDuration, '
-        'transitionCurve: $transitionCurve}';
+        'transitionCurve: $transitionCurve}, '
+        'params: $params';
   }
 }