+++ lib/src/models/sailor_route.dart
@@ -1,10 +1,15 @@
 import 'package:sailor/src/models/base_arguments.dart';
 import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
+import 'package:sailor/src/models/sailor_param.dart';
+import 'package:sailor/src/sailor.dart';
 import 'package:sailor/src/transitions/sailor_transition.dart';
 
 typedef SailorRouteBuilder = Widget Function(
-    BuildContext context, BaseArguments args);
+  BuildContext context,
+  BaseArguments args,
+  ParamMap paramMap,
+);
 
 class SailorRoute {
   final String name;
@@ -13,6 +18,7 @@ class SailorRoute {
   final List<SailorTransition> defaultTransitions;
   final Duration defaultTransitionDuration;
   final Curve defaultTransitionCurve;
+  final List<SailorParam> params;
 
   const SailorRoute({
     @required this.name,
@@ -21,6 +27,7 @@ class SailorRoute {
     this.defaultTransitions,
     this.defaultTransitionDuration,
     this.defaultTransitionCurve,
+    this.params,
   })  : assert(name != null),
         assert(builder != null);
 }